refactor: undate directory structure
