+++ openedx_events/bridge/tests/test_avro_attrs_bridge.py
@@ -7,8 +7,8 @@ from unittest import TestCase
 import attr
 from opaque_keys.edx.keys import CourseKey
 
-from openedx_events.avro_attrs_bridge import AvroAttrsBridge
-from openedx_events.avro_attrs_bridge_extensions import CourseKeyAvroAttrsBridgeExtension
+from openedx_events.bridge.avro_attrs_bridge import AvroAttrsBridge
+from openedx_events.bridge.avro_attrs_bridge_extensions import CourseKeyAvroAttrsBridgeExtension
 from openedx_events.learning.data import CourseData, CourseEnrollmentData, UserData, UserPersonalData
